requirements.txt
anthropic>=0.40.0
pyaudio==0.2.14
SpeechRecognition==3.10.4
pyttsx3==2.90
opencv-python==4.10.0.84
requests==2.32.3
python-dotenv==1.0.1